refactor(media-type.enum): move media-type.enum.ts
Moved media-type.enum.ts to an enum directory and adjusted import paths.
